Add mar12-2 agent results: 7.228 record, hybrid verdict (#25)
New record: 7.228 val_bpb (SSD + compile + seq512, best@3700 of 5000 steps).

Agent results: hybrid (attn@0/2/3 all worse than pure SSD), SwiGLU (32% slower,
same quality), conv_k=8 (worse than k=4), C_SCALE=1.0 (no effect on SSD).
Fix attention mask dtype for bfloat16 compatibility.

README: add leaderboard #6, 7 experiments to details, count 58 to 65.
program.md: priority directive for next run (SSD byte-level, block improvements),
resolved experiments, cleanup (ssd.py editable, SSD run examples, 4000-step optimum).

+++ attention.py
@@ -45,7 +45,7 @@ class AttentionBlock(nn.Module):
             causal = (diff >= 0) & (diff < self.window)
         else:
             causal = diff >= 0
-        mask = mx.where(causal, mx.zeros((L, L)), mx.array(float("-inf")))
+        mask = mx.where(causal, mx.zeros((L, L)), mx.array(float("-inf"))).astype(q.dtype)
 
         scale = 1.0 / (self.d_head**0.5)
         y = mx.fast.scaled_dot_product_attention(q, k, v, scale=scale, mask=mask)
